Moved all the svg image imports to the index file in assets folder
