+++ src/assets/index.js
@@ -1,0 +1,13 @@
+
+import logo from './images/logo.svg'
+import notification from './images/notification.svg'
+import avatar from './images/avatar.svg'
+import search_img from './images/Search-search.svg'
+import cancel from './images/cancel.svg'
+export{
+    logo,
+    notification,
+    avatar,
+    search_img,
+    cancel
+}
